관리자 recommend Controller 세분화
기능 마다 따로 컨트롤러를 세분화 해서 유지 보수에 유용하게 함
