+++ src/main/java/org/flower/controllers/admin/recommend/KeywordController.java
@@ -1,12 +1,9 @@
 package org.flower.controllers.admin.recommend;
 
-import org.flower.entities.Flower;
 import org.flower.entities.Keywords;
-import org.flower.models.recommend.flower.FlowerEditService;
-import org.flower.models.recommend.flower.FlowerInfoService;
+import org.flower.models.recommend.keyword.KeywordEditService;
 import org.flower.models.recommend.keyword.KeywordInfo;
 import org.flower.models.recommend.keyword.KeywordInfoService;
-import org.flower.models.recommend.keyword.KeywordEditService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,40 +16,16 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+/* 관리자 페이지 키워드 추가, 수정, 삭제 관련 컨트롤러 */
 @Controller
 @RequestMapping("/admin/recommend")
-public class RecoMegmentController {
+public class KeywordController {
+
     @Autowired
     private KeywordInfoService keywordInfoService;
 
     @Autowired
     private KeywordEditService keywordEditService;
-
-    @Autowired
-    private FlowerInfoService flowerInfoService;
-
-    @Autowired
-    private FlowerEditService flowerEditService;
-
-    // 꽃 관련 메소드들
-    // 들어가면 바로 꽃 리스트 보이게 하려고 주소 추가로 할당 안함
-    @GetMapping()
-    public String flowerList(Model model){
-        List<Flower> flowersList = flowerInfoService.getAllFlowers();
-        model.addAttribute("flowersList", flowersList);
-        return "admin/recommend/index";
-    }
-    // 꽃 추가 - Post
-    @PostMapping("/addFlower")
-    public String addFlower(@RequestParam String flowerNm, String flowerMean, String bloomseason, String season, String flowerIamges, RedirectAttributes redirectAttributes) {
-        try {
-            flowerEditService.addFlower(flowerNm, flowerMean, bloomseason, season, flowerIamges);
-            redirectAttributes.addFlashAttribute("successMessage", "꽃 성공적으로 추가되었습니다.");
-        }catch (Exception e){
-            redirectAttributes.addFlashAttribute("errorMessage", "꽃 추가 중 오류가 발생했습니다.");
-        }
-        return "redirect:/admin/recommend";
-    }
 
     //키워드 관련 메소드들
     @GetMapping("/keyword")
@@ -63,9 +36,9 @@ public class RecoMegmentController {
     }
 
     /**
-    * 키워드 추가 - POST
-    *
-    * */
+     * 키워드 추가 - POST
+     *
+     * */
     @PostMapping("/addKeyword")
     public String addKeyword(@RequestParam String keywordNm, RedirectAttributes redirectAttributes){
         try {
@@ -78,8 +51,8 @@ public class RecoMegmentController {
     }
 
     /*
-    * 키워드 수정 - POST
-    * */
+     * 키워드 수정 - POST
+     * */
     @PostMapping("/editKeyword")
     public ResponseEntity<Map<String, Object>> editKeyword(@RequestBody KeywordInfo keywordInfo){
         Map<String, Object> response = new HashMap<>();
@@ -97,8 +70,8 @@ public class RecoMegmentController {
     }
 
     /*
-    * 키워드 삭제 - Delete
-    * */
+     * 키워드 삭제 - Delete
+     * */
     @DeleteMapping("/deleteKeywords")
     public ResponseEntity<Map<String, Object>> deleteKeywords(@RequestBody Map<String, List<Long>> payload){
         Map<String, Object> response = new HashMap<>();
@@ -114,11 +87,5 @@ public class RecoMegmentController {
             e.printStackTrace();
             return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
-    }
-
-    @GetMapping("/flowerWeight")
-    public String flowerWeight(){
-
-        return "admin/recommend/flowerWeight";
     }
 }
